some tweaks and kills non injury brutes on carbons

+++ code/modules/spells/spell_types/pointed/attach_bodypart.dm
@@ -30,6 +30,8 @@
 		for(var/datum/wound/limb_wounds as anything in limb.wounds)
 			qdel(limb_wounds)
 		for(var/datum/injury/limb_wounds as anything in limb.injuries)
+			if(limb_wounds.damage_type == WOUND_DIVINE)
+				continue
 			qdel(limb_wounds)
 		limb.update_damages()
 		cast_on.visible_message(
